+++ src/main/kotlin/ru/ac/uniyar/domain/queries/AddProjectQuery.kt
@@ -15,7 +15,7 @@ class AddProjectQuery(store: Store) {
         name: String,
         entrepreneurId: UUID,
         description: String,
-        fundSize: Double,
+        fundSize: Int,
         fundraisingStart: LocalDateTime,
         fundraisingEnd: LocalDateTime,
     ): UUID {
