+++ src/main/kotlin/ru/ac/uniyar/handlers/CreateProjectHandlers.kt
@@ -12,7 +12,7 @@ import org.http4k.lens.Invalid
 import org.http4k.lens.Validator
 import org.http4k.lens.WebForm
 import org.http4k.lens.dateTime
-import org.http4k.lens.double
+import org.http4k.lens.int
 import org.http4k.lens.string
 import org.http4k.lens.uuid
 import org.http4k.lens.webForm
@@ -41,7 +41,7 @@ class AddProjectHandler(
         private val nameFormLens = FormField.string().required("name")
         private val entrepreneurIdFormLens = FormField.uuid().required("entrepreneur")
         private val descriptionFormLens = FormField.string().required("description")
-        private val fundSizeFormLens = FormField.double().required("fundSize")
+        private val fundSizeFormLens = FormField.int().required("fundSize")
         private val fundraisingStartFormLens = FormField.dateTime().required("fundraisingStart")
         private val fundraisingEndFormLens = FormField.dateTime().required("fundraisingEnd")
         private val projectFormLens = Body.webForm(
